qcacld-3.0: Prevent NS offload when active offload is disabled
Currently, Host configures the NS entries through vendor command
where host do not check the active offload is enable or disabled.
Hence it is not configuring according to active offload ini.

As part of fix, Cache the vendor command value and check whether
active offload is enabled or not before configuration to fwr.

Change-Id: Icb11e43198a75975418a2c9ed95d5f1df05ce0c5
CRs-fixed: 2105071